pyproject.toml
[project]
name = "retail-insights-assistant"
version = "1.0.0"
description = "GenAI-powered Retail Insights Assistant with multi-agent NL-to-SQL capabilities"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "Leela Sai Siva D" }
]
keywords = ["nlsql", "langgraph", "duckdb", "genai", "multi-agent"]

dependencies = [
    # Core Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    # LangGraph / LangChain
    "langgraph>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    # Data Layer
    "duckdb>=1.1.0",
    "pandas>=2.2.0",
    "pyarrow>=17.0.0",
    # Database & Cache
    "psycopg[binary,pool]>=3.2.0",
    "redis>=5.0.0",
    # SQL Validation
    "sqlglot>=25.0.0",
    # Utilities
    "cachetools>=5.5.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    # Observability
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]
ui = [
    "streamlit>=1.38.0",
    "plotly>=5.24.0",
    "seaborn>=0.13.0",
    "matplotlib>=3.9.0",
    "openpyxl>=3.1.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]
all = [
    "retail-insights-assistant[dev,ui,notebooks]",
]

[project.scripts]
retail-insights = "retail_insights.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/retail_insights"]

[tool.hatch.build.targets.sdist]
include = ["src"]

# Ruff Configuration
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "F",     # Pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "W",     # pycodestyle warnings
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "UP038",  # Use X | Y in isinstance - requires --unsafe-fixes, keep tuple syntax for readability
]

[tool.ruff.lint.isort]
known-first-party = ["retail_insights"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# MyPy Configuration
[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
ignore_missing_imports = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
no_implicit_optional = false
warn_return_any = false

[[tool.mypy.overrides]]
module = [
    "duckdb.*",
    "langgraph.*",
    "langchain.*",
    "langchain_openai.*",
    "sqlglot.*",
    "streamlit.*",
    "pandas.*",
    "cachetools.*",
]
ignore_missing_imports = true

# Pytest Configuration
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "-v",
    "--tb=short",
    "--cov=src/retail_insights",
    "--cov-report=term-missing",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "benchmark: Performance benchmark tests",
]

# Coverage Configuration
[tool.coverage.run]
source = ["src/retail_insights"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

# Bandit Security Configuration
[tool.bandit]
exclude_dirs = ["tests", ".venv", "notebooks"]
skips = ["B101"]  # Skip assert warnings (used in tests)
targets = ["src/retail_insights"]

[dependency-groups]
dev = [
    "bandit>=1.9.3",
]
